LayoutTests/media/modern-media-controls/fullscreen-support/ipad/fullscreen-support-tap.html

<script src="../../../../resources/js-test-pre.js"></script>
<script src="../../resources/media-controls-loader.js" type="text/javascript"></script>
<script src="../../resources/media-controls-utils.js" type="text/javascript"></script>
<body>
<style type="text/css" media="screen">
    
    video, #host {
        position: absolute;
        top: 0;
        left: 0;
    }

    video {
        width: 768px;
        height: 240px;
    }
    
</style>
<video src="../../../content/test.mp4" autoplay controls></video>
<div id="host"></div>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>FullscreenSupport</code> behavior when tapping the fullscreen button.");

const container = document.querySelector("div#host");
const media = document.querySelector("video");
const mediaController = createControls(container, media, null);

media.addEventListener("webkitpresentationmodechanged", () => {
    if (media.webkitDisplayingFullscreen) {
        shouldBeTrue("media.webkitDisplayingFullscreen");
        media.webkitExitFullscreen();
    } else {
        // We're giving the test time to finish the animation to exit fullscreen
        // so that we are in a clean state for the next test.
        setTimeout(() => {
            container.remove();
            media.remove();
            finishJSTest();
        }, 1000);
    }
});

scheduler.frameDidFire = function() {
    if (media.paused)
        return;

    if (pressOnElement(mediaController.controls.fullscreenButton.element))
        scheduler.frameDidFire = null;
};

</script>
<script src="../../../../resources/js-test-post.js"></script>
</body>
