LayoutTests/media/modern-media-controls/fullscreen-support/ipad/fullscreen-support-tap-expected.txt

Testing the FullscreenSupport behavior when tapping the fullscreen button.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS media.webkitDisplayingFullscreen is true
PASS successfullyParsed is true

TEST COMPLETE
